Homeworks/HW 02/DesignPatterns_HW2/DesignPatterns_HW2.Tests/Transformations/NormalizeTransformationTests.cs
using DesignPatterns_HW2.Transformations;

namespace DesignPatterns_HW2.Tests.Transformations
{
    [TestFixture]
    public class NormalizeTransformationTests
    {
        private NormalizeTransformation _transformation;

        [SetUp]
        public void SetUp()
        {
            _transformation = new NormalizeTransformation();
        }

        [Test]
        public void TransformShouldNormalizedText()
        {
            // Arrange
            var expected = "Hello World!";
            var text = "   Hello   World!   ";

            // Act
            var actual = _transformation.Transform(text);

            // Assert
            Assert.That(actual, Is.EqualTo(expected));
        }

        [Test]
        public void EqualsShouldReturnTrueWhenOtherTransformationIsNormalizeTransformation()
        {
            // Arrange
            var otherTransformation = new NormalizeTransformation();

            // Act
            var actual = _transformation.Equals(otherTransformation);

            // Assert
            Assert.That(actual, Is.True);
        }

        [Test]
        public void EqualsShouldReturnFalseWhenOtherTransformationIsNull()
        {
            // Arrange
            // Act
            var actual = _transformation.Equals(null);

            // Assert
            Assert.That(actual, Is.False);
        }

        [Test]
        public void EqualsShouldReturnFalseWhenOtherObjectIsNotNormalizeTransformation()
        {
            // Arrange
            var otherObj = new object();

            // Act
            var actual = _transformation.Equals(otherObj);

            // Assert
            Assert.That(actual, Is.False);
        }
    }
}
